Errorfix
Fixade lite errors, där scripts krashade med varandra. Det borde var rätt nu.
